セーブ機能強化

+++ Assets/Scripts/Save.cs
@@ -11,38 +11,48 @@ public class Save : MonoBehaviour {
     private int levelshash = 11;//ハッシュキーは11
     private int hpshash = 23;
     public Datamanagement datamanagement;
-	public bool Load(int num)//num番目のセーブデータを読み込む
+    public string named;
+	public string Load(int num)//num番目のセーブデータを読み込む
     {
         if (PlayerPrefs.HasKey("hp"+num))//もし"hp(num)"というデータが存在したら
         {
             if (PlayerPrefs.HasKey("level" + num))//もし"level(num)"というデータが存在したら
             {
-                leveldata = PlayerPrefs.GetInt("level" + num);
-                hpdata = PlayerPrefs.GetInt("hp" + num);
-                levelhash = PlayerPrefs.GetInt("levelh" + num);
-                hphash = PlayerPrefs.GetInt("hph" + num);
-                if (leveldata % levelshash != levelhash)
+                if (PlayerPrefs.HasKey("name" + num))
                 {
-                    PlayerPrefs.DeleteAll();//もし改変されていたらすべてのデータを削除する
-                    return false;
+                    leveldata = PlayerPrefs.GetInt("level" + num);
+                    hpdata = PlayerPrefs.GetInt("hp" + num);
+                    levelhash = PlayerPrefs.GetInt("levelh" + num);
+                    hphash = PlayerPrefs.GetInt("hph" + num);
+                    if (leveldata % levelshash != levelhash)
+                    {
+                        PlayerPrefs.DeleteAll();//もし改変されていたらすべてのデータを削除する
+                        return "save data has breaked!";
+                    }
+                    if (hpdata % hpshash != hphash)
+                    {
+                        PlayerPrefs.DeleteAll();
+                        return "save data has breaked!";
+                    }
+                    if (leveldata % 101 != 0)
+                    {
+                        PlayerPrefs.DeleteAll();
+                        return "save data has breaked!";
+                    }
+                    leveldata = leveldata % 101;
+                    datamanagement.Saver(hpdata, leveldata);
+                    named = PlayerPrefs.GetString("name" + num);
                 }
-                if (hpdata % hpshash != hphash)
-                {
-                    PlayerPrefs.DeleteAll();
-                    return false;
-                }
-                if (leveldata % 101 != 0)
-                {
-                    PlayerPrefs.DeleteAll();
-                }
-                leveldata = leveldata % 101;
-                datamanagement.Saver(hpdata, leveldata);
             }
         }
-        return false;//データがない場合はfalseを返す
+        return "nothing";//データがない場合はfalseを返す
     }
-    public void Saved(int num)
+    public void Saved(int num,string names)
     {
+        if (PlayerPrefs.HasKey("name" + num))
+        {
+            PlayerPrefs.DeleteKey("name" + num);
+        }
         if (PlayerPrefs.HasKey("hp" + num))
         {
             PlayerPrefs.DeleteKey("hp" + num);
@@ -53,9 +63,10 @@ public class Save : MonoBehaviour {
             PlayerPrefs.DeleteKey("level" + num);
             PlayerPrefs.DeleteKey("levelh" + num);
         }
+        PlayerPrefs.SetString("name" + num, names);
         PlayerPrefs.SetInt("hp" + num, (datamanagement.hitpoint)*101);
         PlayerPrefs.SetInt("hph" + num, ((datamanagement.hitpoint*101) % hpshash));
-        PlayerPrefs.SetInt("level" + num, datamanagement.level);
-        PlayerPrefs.SetInt("levelh" + num, (datamanagement.level % levelshash));
+        PlayerPrefs.SetInt("level" + num, (datamanagement.level));
+        PlayerPrefs.SetInt("levelh" + num, ((datamanagement.level) % levelshash));
     }
 }
